% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airDensity.R
\name{airDensity}
\alias{airDensity}
\title{Air density in \link{g cm^-3}}
\usage{
airDensity(Temp, p, h, units = c("deg.C", "hPa", "\%"), opt = "B")
}
\arguments{
\item{Temp}{temperature}
}
\value{
air density value
}
\description{
Calculates the density of the air in the laboratory according to two models that use
temperature, barometric pressure and relative humidity.
}
\examples{
airDensity(Temp = 22.3, p = 748.1, h = 37, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%')) # [g/cm^3]
airDensity(Temp = 23.4, p = 612.3, h = 23, units = c('deg.C', 'mmHg', '\%'), opt = 'A') # [g/cm^3]

}
