% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/df2t.curve.R
\name{df2t.curve}
\alias{df2t.curve}
\title{Converts a dataframe to}
\usage{
df2t.curve(
  df,
  colTit = 1,
  colSignal = 2,
  plot = TRUE,
  smoothLi = TRUE,
  smoothDer1 = TRUE,
  derScaling = 0.6,
  length = 10000,
  main = NULL
)
}
\description{
Converts a dataframe to
}
