% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP.1stDer.R
\name{EP.1stDer}
\alias{EP.1stDer}
\title{Potentiometric equivalence point using first derivate method}
\usage{
EP.1stDer(
  curve,
  length = 10000,
  sub = FALSE,
  subregion = c(0.985, 1.015),
  plot = FALSE,
  ...
)
}
\description{
Potentiometric equivalence point using first derivate method
}
