% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EP.numGrad.R
\name{EP.numGrad}
\alias{EP.numGrad}
\title{Numeric final point in potentiometric indication}
\usage{
EP.numGrad(
  curve,
  length = 10000,
  plot = FALSE,
  upwards = FALSE,
  vabline = TRUE,
  inter = FALSE,
  ...
)
}
\description{
Numeric final point in potentiometric indication
}
