% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lims2t.curve.R
\name{lims2t.curve}
\alias{lims2t.curve}
\title{Converts LIMS reports to data frames}
\usage{
lims2t.curve(
  file,
  skip = 25,
  nrows = 101,
  header = FALSE,
  colTit = 2,
  colSignal = 3,
  plot = TRUE,
  smoothLi = TRUE,
  smoothDer1 = TRUE,
  derScaling = 0.6,
  length = 10000,
  main = NULL
)
}
\description{
Converts LIMS reports to data frames
}
