% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCurve.R
\name{plotCurve}
\alias{plotCurve}
\title{Plots titration curves}
\usage{
plotCurve(
  curve,
  xlab = "Titulante (g)",
  ylab = "Potencial (mV)",
  smoothLi = TRUE,
  smoothDer1 = TRUE,
  smoothDer2 = FALSE,
  length = 10000,
  main = NULL,
  spar = 0.2,
  plot = TRUE,
  vabline = TRUE,
  mn = NULL
)
}
\description{
Plots titration curves
}
